Background-Color-Changer
In this project i practice my skill to change the background of screen display on the the laptop or other devices. In this Project I use HTML5, CSS3 and JAVASCRIPT to complete this project. It is simple but i trying to make more complex project to practice my skills. Please like and comment it will motivate me. Thanks